Restructure library into something suitable for general genart
Closes #48
