+++ generative/lsystem/grammar.py
