+++ generative/lsystem/interpreter.py
