+++ tools/wkt2svg.py
@@ -13,7 +13,7 @@ import shapely.geometry
 
 root = pathlib.Path(__file__).resolve().parent.parent
 sys.path.insert(0, str(root))
-from lsystem import deserialize_geometries  # isort:skip
+from generative.wkio import deserialize_geometries  # isort:skip
 
 LOG_LEVELS = {
     "CRITICAL": logging.CRITICAL,
